% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simevol1.R
\docType{package}
\name{simevol-package}
\alias{simevol-package}
\alias{_PACKAGE}
\alias{simevol}
\title{Simulator of adaptive evolution.}
\description{
Simulator for adaptive evolution.
}
\author{
\strong{Maintainer}: Hiroshi C. Ito \email{hiroshibeetle@gmail.com}

}
\keyword{internal}
